examples/chirpstack-application-server/integrations/azure-service-bus/python/requirements.txt
chirpstack-api==3.7.2
azure-servicebus==0.50.3
